+++ go/internal/install/path_stub.go
@@ -1,0 +1,6 @@
+//go:build !windows
+
+package install
+
+func installWindowsPATH(Options, *Result) error { return nil }
+func removeWindowsPATH(Options, *Result)       {}