+++ go/internal/install/profiles_stub.go
@@ -1,0 +1,6 @@
+//go:build !windows
+
+package install
+
+func installWindowsCompletionHooks(Options, *Result) error { return nil }
+func removeWindowsCompletionHooks(Options, *Result)       {}